[inetstack] Enhancement: Unify config
