+++ src/rust/catpowder/mod.rs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-mod config;
 pub mod runtime;
 
 //==============================================================================
